step 1: move filterby to its own file
