+++ weave/trace_server/trace_server_interface.py
@@ -3,6 +3,7 @@ import datetime
 import typing
 from pydantic import BaseModel
 
+from .interface.filter_by import FilterBy
 
 class CallSchema(BaseModel):
     id: str
@@ -161,83 +162,6 @@ class _CallsFilter(BaseModel):
     wb_run_ids: typing.Optional[typing.List[str]] = None
 
 
-# Can be any standard json-able value
-class _RawValue(BaseModel):
-    value_: typing.Union[
-        str, int, float, bool, dict[str, "_RawValue"], list["_RawValue"]
-    ]
-
-
-# Field should be a key of `CallSchema`. For dictionary fields (`attributes`,
-# `inputs`, `outputs`, `summary`), the field can be dot-separated.
-class _FieldSelect(BaseModel):
-    field_: str
-    # Should this be an operation??
-    cast_: typing.Optional[typing.Literal["str", "int", "float", "bool"]] = None
-
-
-_Operand = typing.Union[_RawValue, _FieldSelect, "_Operation"]
-
-# Operations: all operations have the form of a single property
-# with the name of the operation suffixed with an underscore.
-# Subset of Mongo Operators: https://www.mongodb.com/docs/manual/reference/operator/query/
-# Starting with these operators for now since they are the most common and with negation
-# can cover most of the other operators.
-class _AndOperation(BaseModel):
-    and_: typing.Tuple["_Operand", "_Operand"]
-
-
-class _OrOperation(BaseModel):
-    or_: typing.Tuple["_Operand", "_Operand"]
-
-
-class _NotOperation(BaseModel):
-    not_: "_Operand"
-
-
-class _EqOperation(BaseModel):
-    eq_: typing.Tuple["_Operand", "_Operand"]
-
-
-class _GtOperation(BaseModel):
-    gt_: typing.Tuple["_Operand", "_Operand"]
-
-
-class _GteOperation(BaseModel):
-    gte_: typing.Tuple["_Operand", "_Operand"]
-
-
-class _LikeOperation(BaseModel):
-    like_: typing.Tuple["_Operand", "_Operand"]
-
-
-_Operation = typing.Union[
-    _AndOperation,
-    _OrOperation,
-    _NotOperation,
-    _EqOperation,
-    _GtOperation,
-    _GteOperation,
-    _LikeOperation,
-]
-
-
-# Update the models to include the recursive types
-_RawValue.model_rebuild()
-_FieldSelect.model_rebuild()
-_AndOperation.model_rebuild()
-_OrOperation.model_rebuild()
-_NotOperation.model_rebuild()
-_EqOperation.model_rebuild()
-_GtOperation.model_rebuild()
-_GteOperation.model_rebuild()
-_LikeOperation.model_rebuild()
-
-
-class _FilterBy(BaseModel):
-    filter: _Operation
-
-
 # Field should be a key of `CallSchema`. For dictionary fields (`attributes`,
 # `inputs`, `outputs`, `summary`), the field can be dot-separated.
 class _SortBy(BaseModel):
@@ -254,7 +178,7 @@ class CallsQueryReq(BaseModel):
     # Sort by multiple fields
     sort_by: typing.Optional[typing.List[_SortBy]] = None
     filter_by: typing.Optional[
-        _FilterBy
+        FilterBy
     ] = None  # should this be moved inside of "filter?"
 
 
@@ -266,7 +190,7 @@ class CallsQueryStatsReq(BaseModel):
     project_id: str
     filter: typing.Optional[_CallsFilter] = None
     filter_by: typing.Optional[
-        _FilterBy
+        FilterBy
     ] = None  # should this be moved inside of "filter?"
